refactor(prepush): factorize parsing of prepush hook variables
